// src/model/fsm/fsm.states.ts
/*
 *  __________________
 *  Qaobee
 *  __________________
 *
 *  Copyright (c) 2015.
 *  Qaobee
 *  All Rights Reserved.
 *
 *  NOTICE: All information contained here is, and remains
 *  the property of Qaobee and its suppliers,
 *  if any. The intellectual and technical concepts contained
 *  here are proprietary to Qaobee and its suppliers and may
 *  be covered by U.S. and Foreign Patents, patents in process,
 *  and are protected by trade secret or copyright law.
 *  Dissemination of this information or reproduction of this material
 *  is strictly forbidden unless prior written permission is obtained
 *  from Qaobee.
 */

/**
 * States
 */
export enum FSMStates {
    PAUSED = 'PAUSED',
    INIT = 'INIT',
    GROUND_SELECTED = 'GROUND_SELECTED',
    GOAL_SELECTED = 'GOAL_SELECTED',
    SELECTED_PLAYER = 'SELECTED_PLAYER',
    DEFENSE = 'DEFENSE',
    GAME_STARTED = 'GAME_STARTED',
    ATTACK = 'ATTACK',
    RETURN_TO_SELECTED_PLAYER = 'RETURN_TO_SELECTED_PLAYER',
    GOAL_LEAVED = 'GOAL_LEAVED',
    GROUND_LEAVED = 'GROUND_LEAVED',
    GAME_ENDED = 'GAME_ENDED'
}